feat: Split API calls for renderer from main
Calls from the renderer to main are still handled by registerIpcHandler.
On the other hand, change the call from main to the renderer to be
independent as a function that returns the emitter.

BREAKING CHANGE: change the call from main to the renderer to be independent as a
function that returns the emitter.(getIpcApiEmitter)